data-engineering-on-gcp-main/Dag_siniestro/cloud_function_dev_trigger_parametros/requirements.txt
google-auth
google-auth-httplib2
google-auth-oauthlib
requests
google-cloud-storage
mysql-connector-python
sqlalchemy
pymysql
google-cloud-secret-manager==2.10.0